RichRopes

RichRopes.jl is a package implementing a rope data structure. This is one of the classic string data structures used in text editing, with good performance for insertion, deletion, and concatenation.

RichRopes enrich the nodes of the rope by including counts of Chars, lines, and graphemes. They are immutable, returning a new rope for every operation. This allows for shared structure, and easy-to-implement undo operations: simply return a previous version of the rope.

RichRopes are natively indexed by Char, not bytes, meaning rope[5:10] will return the fifth through tenth character. By using StringUnits, codeunits and graphemes may also be used to index and slice efficiently. Textwidth is not stored in nodes, because this metric is mainly useful for reasonably-short spans such as a single line, but tw units from StringUnits are also supported for RichRope operations.

Support for the AbstractString interface is complete as to the basics. Currently absent are any searching operations, such as findfirst|last or occursin. We intend to support these for literal strings, unfortunately, regex is currently impractical, since PCRE operates on the C interface, not natively Julian data structures.

RichRopes.AbstractRope

RichRopes.cleave

RichRopes.cleave

RichRopes.compactleaves!

RichRopes.delete

RichRopes.delete

RichRopes.grapheme

RichRopes.graphemeindex

RichRopes.nthgraphemeindex

RichRopes.rebuild

RichRopes.splice

RichRopes.AbstractRope — Type

AbstractRope <: AbstractString

Supertype for Rope data structures.

RichRopes.cleave — Method

cleave(rope::RichRope, range::UnitRange{<:Integer})

Return two ropes which are the head and tail of rope with range removed. Range is measured in codepoints.

RichRopes.cleave — Method

cleave(rope::RichRope, index::Integer)

Return two ropes created by splitting rope at index, as indexed by codepoints.

RichRopes.compactleaves! — Method

compactleaves!(leaves::Vector{RichRope{<:AbstractString,Nothing}}, leafsize=leaf_size[])

Compact a Vector of RichRope leaves, such that the returned Vector contains leaves of approximately leafsize.

Note: this will never make a leaf smaller through splitting, so it cannot be used to reduce the leaf size of a RichRope. Presuming that the rope was created originally of leaf_size[] leaves, the leaves resulting from compactleaves! will never be larger than 1.75X leaf_size[].

This is called by rebuild.

RichRopes.delete — Method

delete(rope::RichRope, index::Integer, n::Integer)

Return a rope with n characters removed, beginning with the character at index.

RichRopes.delete — Method

delete(rope::RichRope, range::UnitRange{<:Integer})

Return a rope with the characters in range deleted.

RichRopes.grapheme — Method

grapheme(rope::RichRope{S,RichRope{S}}, index::Integer) where {S<:AbstractString}

Return the indexth grapheme of the rope.

RichRopes.graphemeindex — Method

graphemeindex(rope::RichRope{S,RichRope{S}}, index::Integer) where {S<:AbstractString}

Return the native (character) index of the indexth grapheme.

RichRopes.nthgraphemeindex — Method

nthgraphemeindex(s::S, i::Integer) where {S<:AbstractString}

Return the (character) index of the ith grapheme

RichRopes.rebuild — Function

rebuild(rope::RichRope{S,RichRope{S}} where {S}, leafsize::Integer=leaf_size[])

Rebuild a RichRope. This will compact adjacent small leaves into correctly-sized ones, and balance the tree. Note that a value of leafsize smaller than that used to create the rope will not result in any leaf nodes being split, but will affect the target length of leaves which are compacted (see compactleaves!).

RichRopes.splice — Method

splice(rope::RichRope, at::Union{Integer,UnitRange{<:Integer}}, str::AbstractString)

Return a new rope with str spliced into rope. If at is an Integer, the rope is split at at, if at is a range, that range is removed.

Example

julia> String(splice(RichRope("aabbcc"), 3:4, "!!"))
"aa!!cc"

julia> String(splice(RichRope("aabbcc"), 3, "!!"))
"aab!!bcc"